+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2.1.0](https://github.com/dgrzeszczak/KeyedCodable/releases/tag/2.1.0)
+
+#### New
+- added support for manual coding with keyed AnyKey 
+- Keyed<> struct added 
+
 ## [2.0.0](https://github.com/dgrzeszczak/KeyedCodable/releases/tag/2.0.0)
 
 #### New